fix(evaluation): widen sample-count products in Cochran Q, DeLong, Kruskal-Wallis and Wilcoxon
CodeQL cs/loss-of-precision in src/Evaluation/Statistics: grandTotal^2 (Cochran Q, wrong Q once
total successes pass 46,340), pos*neg (DeLong AUC), N(N+1) and N^3-N (Kruskal-Wallis H, tie
correction wrong from N = 1,291, Dunn post-hoc p NaN from 46,341) and nr(nr+1)(2nr+1)
(Wilcoxon std/p NaN from ~1,291 non-zero pairs) were int products.

+++ src/Evaluation/Statistics/CochranQTest.cs
@@ -85,7 +85,7 @@ public class CochranQTest<T> : IClassifierComparisonTest<T>
         double sumRowSquares = rowTotals.Sum(r => (double)r * r);
         double sumRowTotals = rowTotals.Sum();
 
-        double numerator = (k - 1) * (k * sumColSquares - grandTotal * grandTotal);
+        double numerator = (k - 1) * (k * sumColSquares - (double)grandTotal * grandTotal);
         double denominator = k * sumRowTotals - sumRowSquares;
 
         if (Math.Abs(denominator) < 1e-10)
